Implement L2 eff_config_thermal, bulk_pwr_throttle
    Implemented p9_mss_bulk_pwr_throttles
                p9_mss_utils_to_throttles
                p9_mss_eff_config_thermal
    Updated throttle xml

Original-Change-Id: I0492f9fe8a5bd027ff5875f236711bcf55af88f5
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31804
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>

+++ src/import/chips/p9/procedures/hwp/memory/p9_mss_throttle_mem.C
@@ -53,8 +53,8 @@ extern "C"
 
         for (const auto& l_mca : mss::find_targets<fapi2::TARGET_TYPE_MCA> (i_target))
         {
-            uint32_t l_runtime_port = 0;
-            uint32_t l_runtime_slot = 0;
+            uint16_t l_runtime_port = 0;
+            uint16_t l_runtime_slot = 0;
             uint32_t l_throttle_denominator = 0;
 
             FAPI_TRY( mss::mem_m_dram_clocks( l_mca, l_throttle_denominator) );
